Make agent runtime seams easier to review and extend
The runtime index mixed reusable helper logic with the core AgentRuntime
class, making every runtime change touch the same oversized file.
Extracting the skill-policy and streamed tool-result helpers into their
own modules keeps the public entrypoint stable while shrinking the next
split surface.

Constraint: Keep src/agent/runtime/index.ts behavior stable while task 44 narrows the split incrementally
Constraint: Avoid overlap with adjacent open work such as PR #1103
Rejected: Continue extracting runtime-state/model-transport helpers in the same pass | broader risk than needed for a reviewable first PR
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Preserve index-level re-exports while splitting runtime helpers so downstream imports remain stable until the larger runtime/assembly split finishes
Tested: deno test --no-check --allow-all src/agent/runtime/skill-policy.test.ts src/agent/runtime/index.test.ts; deno check src/agent/runtime/index.ts
Not-tested: Full veryfront-code test suite; repo-wide pre-commit verification was not re-run after generated artifact cleanup
